Finished navbar and added styling to other bits Added routing so the navbar works, successfully linking to each screen. Stats, Tips and Urgent Support pages will need some styling to get the navbar etc back to the top of the page (currently in the middle because of justify-content:center inherited from the fullscreen landing page).
Added styling to buttons on homepage so they go lighter on hovering.

Added styling to emphasis urgent support button.

Changed font colour to the default navy for placeholder text on
dashboard.

Added a max width for the dashboard content so it didn't stretch so
wide. Probably won't make a difference when it's replaced with actual
blurb though because it won't be as much text.

Added Facts.js component and Facts.css to put the placeholder text in a
separate component to <Home />.